move events to its own route
